test(resource|domain): tags create read and delete test steps

+++ monte_carlo/client/monte_carlo_client.go
@@ -158,6 +158,6 @@ type GetDomain struct {
 
 type DeleteDomain struct {
 	DeleteDomain struct {
-		Deleted uint64
+		Deleted int
 	} `graphql:"deleteDomain(uuid: $uuid)"`
 }
